Rotating around object now seems to work. The next step is to make it work when the charcter is outside the bounding box. Currrently it only works iof the charcter does not leave the bounding box.

+++ actors/Player.hpp
@@ -21,7 +21,7 @@ class Player: public GameObject
         Sprite *sprite;
         Shape line;
 
-        Vector2i* rayTrace(Sprite *sprite, int fromX, int fromY, int toX, int toY);
+        Vector2f* rayTrace(Sprite *sprite, int fromX, int fromY, int toX, int toY);
 };
 
 #endif // PLAYER_H
